feat: moved RoomService RoomTypeService to impl package and added 'Impl' at the end of the class names. Added interfaces they implement to match project structure
